data dari bulus

+++ resources/views/ppmb/jadwal/detjadwal/detjadwal_lihat.blade.php
@@ -56,7 +56,7 @@
 										<i class="material-icons">Edit</i><span class="icon-name"></span>
 									</button>
 								</a>
-								<a href="/detjadwal_ppmb/delete/{{$data->id_detjadwal}}">
+								<a href="/detjadwal_ppmb/delete/{{$data->id_detjadwal}}" onclick="return confirm('Hapus data ?');">
 									<button type="button" class="btn btn-danger btn-square waves-effect waves-square waves-effect">
 										<i class="material-icons">Hapus</i><span class="icon-name"></span>
 									</button>
