+++ resources/views/ppmb/jadwal/lihat_jadwal.blade.php
@@ -38,7 +38,7 @@
 										<i class="material-icons">Edit</i><span class="icon-name"></span>
 									</button>
 								</a>
-								<a href="/jadwal_ppmb/delete/{{$data->id_jadwal}}">
+								<a href="/jadwal_ppmb/delete/{{$data->id_jadwal}}" onclick="return confirm('Hapus data ?');">
 									<button type="button" class="btn btn-danger btn-square 
 									waves-effect waves-square waves-effect">
 									<i class="material-icons">Hapus</i><span class="icon-name"></span>
